+++ libs/typescript/playground/src/styles.css
@@ -1,0 +1,121 @@
+@import "tailwindcss";
+
+/* Standalone dark mode - works with both class and media query */
+@custom-variant dark (&:where(.dark, .dark *));
+
+@theme {
+  /* Custom theme colors */
+  --color-secondary-50: oklch(0.97 0.014 254.604);
+  --color-secondary-100: oklch(0.932 0.032 255.585);
+  --color-secondary-200: oklch(0.882 0.059 254.128);
+  --color-secondary-300: oklch(0.809 0.105 251.813);
+  --color-secondary-400: oklch(0.707 0.165 254.624);
+  --color-secondary-500: oklch(0.623 0.214 259.815);
+  --color-secondary-600: oklch(0.546 0.245 262.881);
+  --color-secondary-700: oklch(0.488 0.243 264.376);
+  --color-secondary-800: oklch(0.424 0.199 265.638);
+  --color-secondary-900: oklch(0.379 0.146 265.522);
+  --color-secondary-950: oklch(0.282 0.091 267.935);
+
+  /* Override blue with sky */
+  --color-blue-50: oklch(0.977 0.013 236.62);
+  --color-blue-100: oklch(0.951 0.026 236.824);
+  --color-blue-200: oklch(0.901 0.058 230.902);
+  --color-blue-300: oklch(0.828 0.111 230.318);
+  --color-blue-400: oklch(0.746 0.16 232.661);
+  --color-blue-500: oklch(0.685 0.169 237.323);
+  --color-blue-600: oklch(0.588 0.158 241.966);
+  --color-blue-700: oklch(0.5 0.134 242.749);
+  --color-blue-800: oklch(0.443 0.11 240.79);
+  --color-blue-900: oklch(0.391 0.09 240.876);
+  --color-blue-950: oklch(0.293 0.066 243.157);
+}
+
+@theme inline {
+  --radius-sm: calc(var(--radius) - 4px);
+  --radius-md: calc(var(--radius) - 2px);
+  --radius-lg: var(--radius);
+  --radius-xl: calc(var(--radius) + 4px);
+  --color-background: var(--background);
+  --color-foreground: var(--foreground);
+  --color-card: var(--card);
+  --color-card-foreground: var(--card-foreground);
+  --color-popover: var(--popover);
+  --color-popover-foreground: var(--popover-foreground);
+  --color-primary: var(--primary);
+  --color-primary-foreground: var(--primary-foreground);
+  --color-secondary: var(--secondary);
+  --color-secondary-foreground: var(--secondary-foreground);
+  --color-muted: var(--muted);
+  --color-muted-foreground: var(--muted-foreground);
+  --color-accent: var(--accent);
+  --color-accent-foreground: var(--accent-foreground);
+  --color-destructive: var(--destructive);
+  --color-border: var(--border);
+  --color-input: var(--input);
+  --color-ring: var(--ring);
+  --color-chart-1: var(--chart-1);
+  --color-chart-2: var(--chart-2);
+  --color-chart-3: var(--chart-3);
+  --color-chart-4: var(--chart-4);
+  --color-chart-5: var(--chart-5);
+}
+
+:root {
+  --radius: 0.625rem;
+  --background: oklch(0.985 0 0);
+  --foreground: oklch(0.145 0 0);
+  --card: oklch(1 0 0);
+  --card-foreground: oklch(0.145 0 0);
+  --popover: oklch(1 0 0);
+  --popover-foreground: oklch(0.145 0 0);
+  --primary: oklch(0.205 0 0);
+  --primary-foreground: oklch(0.985 0 0);
+  --secondary: oklch(0.97 0 0);
+  --secondary-foreground: oklch(0.205 0 0);
+  --muted: oklch(0.97 0 0);
+  --muted-foreground: oklch(0.556 0 0);
+  --accent: oklch(0.97 0 0);
+  --accent-foreground: oklch(0.205 0 0);
+  --destructive: oklch(0.577 0.245 27.325);
+  --border: oklch(0.922 0 0);
+  --input: oklch(0.922 0 0);
+  --ring: oklch(0.708 0 0);
+  --chart-1: oklch(0.646 0.222 41.116);
+  --chart-2: oklch(0.6 0.118 184.704);
+  --chart-3: oklch(0.398 0.07 227.392);
+  --chart-4: oklch(0.828 0.189 84.429);
+  --chart-5: oklch(0.769 0.188 70.08);
+}
+
+.dark {
+  --background: oklch(0.09 0 0);
+  --foreground: oklch(0.985 0 0);
+  --card: oklch(0.205 0 0);
+  --card-foreground: oklch(0.985 0 0);
+  --popover: oklch(0.205 0 0);
+  --popover-foreground: oklch(0.985 0 0);
+  --primary: oklch(0.922 0 0);
+  --primary-foreground: oklch(0.205 0 0);
+  --secondary: oklch(0.269 0 0);
+  --secondary-foreground: oklch(0.985 0 0);
+  --muted: oklch(0.269 0 0);
+  --muted-foreground: oklch(0.708 0 0);
+  --accent: oklch(0.269 0 0);
+  --accent-foreground: oklch(0.985 0 0);
+  --destructive: oklch(0.704 0.191 22.216);
+  --border: oklch(1 0 0 / 10%);
+  --input: oklch(1 0 0 / 15%);
+  --ring: oklch(0.556 0 0);
+  --chart-1: oklch(0.488 0.243 264.376);
+  --chart-2: oklch(0.696 0.17 162.48);
+  --chart-3: oklch(0.769 0.188 70.08);
+  --chart-4: oklch(0.627 0.265 303.9);
+  --chart-5: oklch(0.645 0.246 16.439);
+}
+
+@layer base {
+  * {
+    @apply border-border outline-ring/50;
+  }
+}
